+++ NEWS.md
@@ -4,6 +4,7 @@
 * Added new measures `classif.bbrier` (binary Brier score) and `classif.mbrier`
   (multi-class Brier score).
 * Added new Resampling: `ResamplingInsample`.
+* Added base class for unsupervised tasks: `TaskUnsupervised`.
 
 # mlr3 0.1.8